Allow gdal to throw exceptions
By default, gdal python binding doesn't throw exceptions, which is the behavior we relied on.
However, another project allows gdal binding to throw an exception causing functions in this libraries fail. Since it's commonly a better thing to handle exceptions than ignore exceptions, we allow gdal to throw exceptions in this library as well.